pyproject.toml
[tool.poetry]
name = "backend-homework"
version = "0.1.0"
description = ""
authors = ["Alejandro Amar Gil <alejandroamar66@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi-login = {extras = ["jwt"], version = "^1.10.0"}
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
loguru = "^0.7.2"
pydantic = {extras = ["email"], version = "^2.7.1"}
pydantic-settings = "^2.2.1"
uvicorn = "^0.29.0"
uvloop = "^0.19.0"
gunicorn = "^22.0.0"
passlib = "^1.7.4"
strawberry-graphql = {extras = ["fastapi"], version = "^0.229.0"}
aiohttp = "^3.9.5"
httpx = "^0.27.0"
pytest = "^8.2.0"
mypy = "^1.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-sugar = "*"  
pytest-cov = "*"  
black = "*"  
mypy = "*"  
ruff = "*"  
isort = "*" 


[tool.pytest.ini_options]
testpaths = ["tests", "integration"]
# addopts = "-v -ra -q"
log_cli = true
log_cli_level = "INFO"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
minversion = "6.0"
filterwarnings = "ignore"
# Specify a custom directory for pytest cache
cache_dir = ".pytest_cache"
addopts = [
    "--import-mode=importlib",
]